pyproject.toml
[project]
name = "genai-launchpad"
version = "1.0.0"
description = "GenAI Launchpad by Datalumina"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "beautifulsoup4>=4.12.2",
    "celery>=5.5.1",
    "fastapi>=0.115.12",
    "firecrawl-py>=1.4.0",
    "instructor>=1.7.9",
    "jinja2>=3.1.6",
    "langfuse==2.60.8",
    "logfire>=3.19.0",
    "nest-asyncio>=1.6.0",
    "opentelemetry-api>=1.33.0",
    "opentelemetry-exporter-otlp>=1.33.0",
    "pandas>=2.3.0",
    "psutil>=5.9.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pydantic-ai[logfire]>=0.1.5",
    "python-dotenv>=1.1.0",
    "python-frontmatter>=1.1.0",
    "redis>=5.2.1",
    "requests>=2.31.0",
    "simple-salesforce>=1.12.5",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "graphviz>=0.20.3",
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "ruff>=0.11.9",
    "nest-asyncio>=1.6.0",
]

[tool.ruff]
exclude = [".venv", "app/core/commands", "app/alembic", ".sh"]

[project.scripts]
createworkflow = "app.core.commands.init_workflow:main"

[tool.setuptools]
packages = ["app"]
